+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   * an `--include-rgb` option has been added to `rtc_sentinel.py`
   * an `include_rgb` keyword argument has been added to `hyp3_gamma.rtc.rtc_sentinel.rtc_sentinel_gamma`
 
+### Changed
+* Upgraded to hyp3lib [v1.6.6](https://github.com/ASFHyP3/hyp3-lib/blob/develop/CHANGELOG.md#166) from v1.6.5
+
 ## [4.0.5](https://github.com/ASFHyP3/hyp3-gamma/compare/v4.0.4...v4.0.5)
 
 ### Changed